fix: HTTP response code "Payload too large"
It *looks* like something outdated again, but I'm still confused with
this.

MDN docs only reference "Payload too large":
https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/413

In ruby stdlib documentation I find both the two constants
`HTTPRequestEntityTooLarge` and `Net::HTTPPayloadTooLarge`.
https://ruby-doc.org/stdlib-2.6.3/libdoc/net/http/rdoc/Net/HTTPPayloadTooLarge.html

Entering 'Net::HTTPPayloadTooLarge' on the ruby Github repository gives
me:
https://github.com/ruby/ruby/blob/48b94b791997881929c739c64f95ac30f3fd0bb9/lib/net/http/responses.rb#L144

The respective commit that changed it:
https://github.com/ruby/ruby/commit/660740a75dd9526ad9a2732ea87c8ab5a385ef1f

The commit message is not very telling and so I am guessing it's a simple
renaming to match the title of the HTTP 413 status code.

This commit fixes one test case on my machine.