refactor: update almost all development deps
I only didn't update `codeclimate-test-reporter` as the gem is
deprecated altogether and I haven't looked into CI pipeline and
understood what's going on.

See: https://github.com/codeclimate/ruby-test-reporter

What was the reason to lift the version constraint on the development
dependencies? Well, there was no version constraint on the runtime
dependencies. I think, a version constraint on the runtime dependencies
would make more sense than on the development dependencies. So I decided
to remove them for the sake of consistency.